# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DJERDJERA Comptable API - Algerian small-business accounting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/items:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items with current balance
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryItem" }
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryItemInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
  /inventory/items/{id}:
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: List stock movements (in/out)
      parameters:
        - name: itemId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryMovement" }
    post:
      operationId: createInventoryMovement
      tags: [inventory]
      summary: Record a stock movement (consumption, production)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InventoryMovementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryMovement" }

  # ============ EMPLOYEES ============
...
  /legal/alerts:
    get:
      operationId: listLegalAlerts
      tags: [legal]
      summary: Journal Officiel texts judged relevant to this company
      description: >
        Newest first. Acknowledged alerts are hidden by default — a business owner
        who has dealt with a text should not have to keep scrolling past it.
      parameters:
        - name: includeAcknowledged
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LegalAlert" }
  /legal/alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeLegalAlert
      tags: [legal]
      summary: Mark a legal alert as dealt with
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalAlert" }
...
components:
...
  schemas:
...
    InventoryItem:
      type: object
      required: [id, name, category, balance, averageCost, totalValue]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        category: { type: string, enum: [RAW_MATERIAL, FINISHED_GOOD, SUPPLY] }
        unit: { type: string, nullable: true, description: "e.g. kg, L, unit" }
        balance: { type: number }
        averageCost: { type: number }
        totalValue: { type: number }
    InventoryItemInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, enum: [RAW_MATERIAL, FINISHED_GOOD, SUPPLY] }
        unit: { type: string, nullable: true }
    InventoryMovement:
      type: object
      required: [id, itemId, itemName, date, quantity, direction, unitCostHt]
      properties:
        id: { type: string, format: uuid }
        itemId: { type: string, format: uuid }
        itemName: { type: string }
        date: { type: string, format: date }
        quantity: { type: number }
        direction: { type: string, enum: [IN, OUT] }
        unitCostHt: { type: number }
        note: { type: string, nullable: true }
        transactionId:
          type: string
          format: uuid
          nullable: true
          description: >-
            The journal entry that generated this movement, if any. Null for
            movements entered by hand on the Stocks page.
    InventoryMovementInput:
      type: object
      required: [itemId, date, quantity, direction]
      properties:
        itemId: { type: string, format: uuid }
        date: { type: string, format: date }
        quantity: { type: number, minimum: 0 }
        direction: { type: string, enum: [IN, OUT] }
        unitCostHt:
          type: number
          nullable: true
          minimum: 0
          description: >-
            Required for an IN. Optional for an OUT, where omitting it (or
            sending 0) applies the article's current CUMP.
        note: { type: string, nullable: true }
...
    LegalAlertMatch:
      type: object
      description: >
        The signals that made this text relevant. Structured rather than prose so
        the client can phrase the reason in the user's own words — a business
        owner needs to see *why* a text concerns them, in a sentence they can
        read, not an opaque score.
      required: [sectors, keywords, legalForm, taxRegime]
      properties:
        sectors: { type: array, items: { type: string } }
        keywords: { type: array, items: { type: string } }
        legalForm: { type: boolean }
        taxRegime: { type: boolean }
    LegalAlert:
      type: object
      required: [id, documentId, docKind, relevance, matchScore, matchedOn, year, issueNumber, createdAt]
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        docKind: { type: string, description: "e.g. Décret exécutif, Arrêté interministériel" }
        docNumber: { type: string, nullable: true }
        titleFr: { type: string, nullable: true }
        titleAr: { type: string, nullable: true }
        summaryFr: { type: string, nullable: true, description: "One sentence, written for a non-lawyer" }
        summaryAr: { type: string, nullable: true }
        publishedOn: { type: string, format: date, nullable: true }
        pageFrom: { type: integer, nullable: true }
        year: { type: integer }
        issueNumber: { type: integer, description: "Journal Officiel issue number" }
        relevance: { type: string, enum: [HIGH, MEDIUM, LOW] }
        matchScore: { type: number }
        matchedOn: { $ref: "#/components/schemas/LegalAlertMatch" }
        acknowledgedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
...